cleanup - added links to home page images
